Adapt test to newer Catalog
ahc component has been removed from Catalog and several tests were
relying on it. They have been adapted:
- some are using timer component instead. A very stable component with
very low chance to become deprecated in the near future
- some are using a specific hand-crafted catalog to be able to test very
specific attributes combinations (consumer only, producer only, both).
the downside is that we are not testing with a "real" component provided
by the catalog. The advantages is that it will be a lot easier to have a
stable test for these very specific cases
- some are using bean component instead because timer can be confusing
for some specific test case as the timer component has a timer attribute
(yes, same name)

Signed-off-by: Aurélien Pupier <apupier@redhat.com>